pyproject.toml
[project]
name = "delete-spam-discord"
version = "0.1.0"
description = "Discord bot for deleting messages based on flexible filter conditions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord.py>=2.3.0",
    "loguru>=0.7.0",
]
